Create readability.c
Here, we check the readability situation of the given string according to a formula. This is the formula; index = 0.0588 * L - 0.296 * S - 15.8. Here, L is the average number of letters per 100 words in the text, and S is the average number of sentences per 100 words in the text. To achieve this calculation, we get letter number, word number and sentence number of a string respectively. Then we put these numbers into a formula. Finally, we find the grade of this string. Here, we use a special header file that is called "cs50.h".